Fix (Compras): Excel sin datos neto/IVA y con decimales
OrdenCompraRepository::getAll no retornaba monto_neto, impuesto ni
campos monetarios. sheetCompras escribía decimales MySQL directo.
Corregido getAll con campos completos y cast a (int) en el export.

+++ insuorders_private/src/App/Controllers/ExportController.php
@@ -390,11 +390,11 @@ class ExportController
                 $d['numero_cotizacion'] ?? '',
                 $d['destino'] ?? '',
                 $d['moneda'] ?? 'CLP',
-                $d['tipo_cambio'] ?? 1,
-                $d['monto_neto'] ?? 0,
+                (int) ($d['tipo_cambio'] ?? 1),
+                (int) ($d['monto_neto'] ?? 0),
                 ($d['impuesto_porcentaje'] ?? 19) . '%',
-                $d['impuesto'] ?? 0,
-                $d['monto_total'] ?? 0,
+                (int) ($d['impuesto'] ?? 0),
+                (int) ($d['monto_total'] ?? 0),
                 $d['creador'] ?? 'Sistema'
             ]
         );
